Remove prys
