+++ lib/app/models/testing_instance.rb
@@ -41,6 +41,9 @@ module Inferno
 
       property :standalone_launch_script, String
 
+      property :initiate_login_uri, String
+      property :redirect_uris, String
+
       has n, :sequence_results
       has n, :supported_resources, order: [:index.asc]
       has n, :resource_references
