+++ lib/app/sequences/02_dynamic_registration_sequence.rb
@@ -59,14 +59,11 @@ module Inferno
 
         params = {
             'client_name' => @instance.client_name,
-            'initiate_login_uri' => "#{@instance.base_url}#{BASE_PATH}/#{@instance.id}/#{@instance.client_endpoint_key}/launch",
-            'redirect_uris' => ["#{@instance.base_url}#{BASE_PATH}/#{@instance.id}/#{@instance.client_endpoint_key}/redirect"],
+            'initiate_login_uri' => @instance.initiate_login_uri,
+            'redirect_uris' => [@instance.redirect_uris],
             'grant_types' => ['authorization_code'],
             'scope' => @instance.scopes,
         }
-
-        params['initiate_login_uri'] = 'http://localhost:4568' if @instance.standalone_launch_script
-        params['redirect_uris'] = ['http://localhost:4568'] if @instance.standalone_launch_script
 
         params['token_endpoint_auth_method'] = if @instance.confidential_client
                                                  'client_secret_basic'
